EXOJCR-1270: clean components before restoring

+++ exo.jcr.component.core/src/main/java/org/exoplatform/services/jcr/impl/backup/rdbms/DirectoryRestor.java
@@ -101,11 +101,14 @@ public class DirectoryRestor implements DataRestor
          try
          {
             File tmpDir = new File(tempDir, PREFIX + IdGenerator.generate());
-            DirectoryHelper.copyDirectory(dataDir, tmpDir);
-
+            PrivilegedFileHelper.mkdirs(tmpDir);
             tmpDirs.add(tmpDir);
 
-            DirectoryHelper.removeDirectory(dataDir);
+            if (PrivilegedFileHelper.exists(dataDir))
+            {
+               DirectoryHelper.copyDirectory(dataDir, tmpDir);
+               DirectoryHelper.removeDirectory(dataDir);
+            }
          }
          catch (IOException e)
          {
